ComputerVision

+++ video/ComputerVision/OpticalFlowLK.cpp
@@ -5,7 +5,7 @@ NAMESPACE_TOPSIDE_BEGIN
 
 
 void optical_flow_lk::track(const pyra8& prev_pyr, pyra8& curr_pyr,
-	const Vector<int>& prev_xy, Vector<double>& curr_xy, int count,
+	const Vector<float>& prev_xy, Vector<float>& curr_xy, int count,
 	int win_size, int max_iter, Vector<uint8>& status,
 	double eps, double min_eigen_threshold) {
 	if (max_iter <= 0) max_iter = 30;
